refactor: organized session.rs
